Fixed clang-tidy errors in search and TT

+++ src/types.h
@@ -13,8 +13,8 @@ namespace sagittar {
     using u64  = std::uint64_t;
     using u128 = unsigned __int128;
 
-    using Score = i16;
-    using Depth = i8;
+    using Score = i32;
+    using Depth = i32;
 
     constexpr int MOVES_MAX = 256;
